updated sub-process function to handle floats

+++ scheduling.h
@@ -32,4 +32,4 @@ int inputProcesses;
 void step(struct cpu *processor, int *processesCompleted);
 void printRunning(struct process processEntry, int id);
 int calculateSplitCount(int time);
-int calculateSubTime(int time, int numberOfSplits);
+int calculateSubTime(float time, float numberOfSplits);
